wip - move NativeSuccess logic into hook
