+++ src/context/WalletProvider/NativeWallet/NativeSuccess/hooks/useNativeSuccess/useNativeSuccess.ts
@@ -1,15 +1,13 @@
-import { ModalBody, ModalHeader, Spinner } from '@chakra-ui/react'
-import { NativeAdapter } from '@shapeshiftoss/hdwallet-native'
-import { useEffect, useState } from 'react'
-import { Card } from 'components/Card/Card'
-import { Text } from 'components/Text'
 import { useLocalStorage } from 'hooks/useLocalStorage/useLocalStorage'
+import { useEffect, useState } from 'react'
+import { useWallet, WalletActions } from 'context/WalletProvider/WalletProvider'
+import { SUPPORTED_WALLETS } from 'context/WalletProvider/config'
+import { NativeSetupProps } from 'context/WalletProvider/NativeWallet/setup'
+import { NativeAdapter } from '@shapeshiftoss/hdwallet-native'
 
-import { SUPPORTED_WALLETS } from '../config'
-import { useWallet, WalletActions } from '../WalletProvider'
-import { NativeSetupProps } from './setup'
+export type UseNativeSuccessPropTypes = Partial<NativeSetupProps>
 
-export const NativeSuccess = ({ location }: NativeSetupProps) => {
+export const useNativeSuccess = ({ location }: NativeSetupProps) => {
   const [isSuccessful, setIsSuccessful] = useState<boolean | null>(null)
   const [, setLocalStorageWallet] = useLocalStorage<Record<string, string>>('wallet', null)
   const { state, dispatch } = useWallet()
@@ -44,24 +42,5 @@ export const NativeSuccess = ({ location }: NativeSetupProps) => {
     })()
   }, [dispatch, location.state, setLocalStorageWallet, state.adapters])
 
-  return (
-    <>
-      <ModalHeader>
-        <Text translation={'walletProvider.shapeShift.nativeSuccess.header'} />
-      </ModalHeader>
-      <ModalBody>
-        <Card mb={4}>
-          <Card.Body fontSize='sm'>
-            {isSuccessful === true ? (
-              <Text translation={'walletProvider.shapeShift.nativeSuccess.success'} />
-            ) : isSuccessful === false ? (
-              <Text translation={'walletProvider.shapeShift.nativeSuccess.error'} />
-            ) : (
-              <Spinner />
-            )}
-          </Card.Body>
-        </Card>
-      </ModalBody>
-    </>
-  )
+  return { isSuccessful }
 }
